Grow awsm.nim, add unit test
Integrated suggestions from elcritch:
https://forum.nim-lang.org/t/13326
Added the all-transitions unit test.

Made progress, but I reached this stopping point.
When a proc is converted by toEventHandler(),
a new instance is created which has a different address;
and we lose the ability to test for equality.
This means we can't detect the converted states
and the framework's init() method fails to detect the
top() state and that breaks the fundamental operation
of the framework.

We either need to find a way to perform type conversion
without creating an object at a new address;
or we need to find a way to refer to the new object at the
new address knowing only the original EventHandler proc.