Get unit tests working
Using unittest2 in order to enable nim-vscode's Test Runner.
I also had to set `"nim.test.entryPoint"` in vscode's settings.json.
The "noNimblePath" was blocking the compiler from finding unittest2,
which was installed by nimble; so I removed nimble.paths.

Enabled initial transition test.  This is the first test that shows
the framework is actually doing what it is designed to do.