Fix transition to current test
The test for transition to current was using the wrong state.
In state 's', the I event will be handled in-state, not transitioned.
We need to use state 's1' and event A to cause a transition
to self that will activate the Exit and Entry actions of 's1'.

I also added the inverse test:
State 's' receiving event I should NOT activate the Exit
and Entry actions of 's'.